backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==3.2.2
python-dotenv==1.0.0
cors==1.0.1

# Blockchain & Web3
web3==6.11.3
eth-account==0.10.0
eth-keys==0.4.0
eth-typing==4.1.0
pysha3==1.0.2

# Encryption & Cryptography
cryptography==42.0.8
pycryptodome==3.19.0
# NOTE: charm-crypto packaging varies by platform; keep it unpinned to avoid install failures.
charm-crypto

# Database
sqlite==0.0.1

# Testing
pytest==7.4.3
httpx==0.25.2

# Utilities
requests==2.31.0

# Cloud storage (GCS) (optional)
# google-cloud-storage==2.14.0

# MongoDB (optional)
pymongo[srv]==4.6.3
